Moved to a submodule and added Data TCK runner.
Provide EntityManager using a CDI provider instead of configuration.
