+++ impl/src/main/java/ee/omnifish/jnosql/jakartapersistence/communication/PersistenceClassScanner.java
