+++ impl/src/main/resources/META-INF/services/org.eclipse.jnosql.mapping.metadata.ClassScanner
