+++ impl/src/test/java/ee/omnifish/jnosql/jakartapersistence/EntityManagerProducer.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2024 OmniFish. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0, which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the
+ * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+ * version 2 with the GNU Classpath Exception, which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ */
+package ee.omnifish.jnosql.jakartapersistence;
+
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.inject.Disposes;
+import jakarta.enterprise.inject.Produces;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.Persistence;
+
+/**
+ *
+ * @author Ondro Mihalyi
+ */
+@ApplicationScoped
+public class EntityManagerProducer {
+    @Produces
+    @ApplicationScoped
+    public EntityManager createEntityManager() {
+        return Persistence.createEntityManagerFactory("testPersistenceUnit")
+                .createEntityManager();
+    }
+
+    public void closeEntityManager(@Disposes EntityManager entityManager) {
+        entityManager.close();
+    }
+}
