+++ impl/src/test/java/ee/omnifish/jnosql/jakartapersistence/Person.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2024 OmniFish
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0, which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the
+ * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+ * version 2 with the GNU Classpath Exception, which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ */
+package ee.omnifish.jnosql.jakartapersistence;
+
+import static jakarta.persistence.GenerationType.AUTO;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import java.util.List;
+
+@Entity(name = "Person")
+public class Person {
+
+  @Id
+  @GeneratedValue(strategy = AUTO)
+  private long id;
+
+  @Column
+  private String name;
+
+  @Column
+  private List<String> phones;
+}
+
