--- src/main/java/ee/omnifish/jnosql/jakartapersistence/communication/PersistenceManagerFactory.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2024 OmniFish. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-package ee.omnifish.jnosql.jakartapersistence.communication;
-
-import jakarta.persistence.EntityManagerFactory;
-
-/**
- *
- * @author Ondro Mihalyi
- */
-public class PersistenceManagerFactory {
-
-    private EntityManagerFactory emf;
-
-    public PersistenceManagerFactory(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-
-    public void close() {
-        emf.close();
-    }
-
-    public PersistenceDatabaseManager apply(String t) {
-        return new PersistenceDatabaseManager(emf.createEntityManager(), emf.getName());
-    }
-
-}
