--- src/test/resources/META-INF/persistence.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- * Copyright (c) 2024 OmniFish. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-  -->
-<persistence xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd" version="2.0">
-  <persistence-unit name="testPersistenceUnit" transaction-type="RESOURCE_LOCAL">
-    <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-    <class>ee.omnifish.jnosql.jakartapersistence.Person</class>
-    <exclude-unlisted-classes>true</exclude-unlisted-classes>
-    <properties>
-      <!-- Common properties -->
-      <property name="jakarta.persistence.jdbc.driver" value="org.apache.derby.jdbc.ClientDataSource"/>
-      <property name="jakarta.persistence.jdbc.url" value="jdbc:derby:memory:target/derbydb/test-jpa;create=true"/>
-      <property name="jakarta.persistence.jdbc.user" value="APP"/>
-      <property name="jakarta.persistence.jdbc.password" value="APP"/>
-
-      <!-- EclipseLink specific properties -->
-      <property name="eclipselink.target-database" value="Derby"/>
-      <property name="eclipselink.ddl-generation" value="drop-and-create-tables"/>
-      <property name="eclipselink.debug" value="ALL"/>
-      <property name="eclipselink.weaving" value="static"/>
-      <property name="eclipselink.logging.level" value="FINEST"/>
-      <property name="eclipselink.logging.level.sql" value="FINEST"/>
-      <property name="eclipselink.logging.level.cache" value="FINEST"/>
-    </properties>
-  </persistence-unit>
-</persistence>
