+++ tck-runner/src/main/resources/META-INF/microprofile-config.properties
