+++ tck-runner/src/main/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,0 +1,1 @@
+ee.omnifish.jakarta.data.tck.runner.CDIExtension
